feat: add responsive get-kali cards
